+++ app/src/main/java/com/fazemeright/chatbotmetcs622/repositories/message/MessageRepositoryImpl.kt
@@ -7,14 +7,12 @@ import com.fazemeright.chatbotmetcs622.database.ChatBotDatabase
 import com.fazemeright.chatbotmetcs622.database.message.Message
 import com.fazemeright.chatbotmetcs622.database.message.Message.Companion.newMessage
 import com.fazemeright.chatbotmetcs622.models.ChatRoom
-import com.fazemeright.library.api.Storable
 import com.fazemeright.library.api.domain.authentication.UserAuthentication
 import com.fazemeright.library.api.domain.authentication.firebase.FireBaseUserAuthentication
 import com.fazemeright.library.api.domain.database.DatabaseStore
 import com.fazemeright.library.api.domain.database.firebase.FireBaseDatabaseStore
 import com.fazemeright.library.api.result.Result
 import kotlinx.coroutines.*
-import kotlinx.coroutines.tasks.await
 import timber.log.Timber
 import java.util.*
 
@@ -36,48 +34,6 @@ class MessageRepositoryImpl private constructor(
             database.messageDao().getLatestMessage(newMessage.chatRoomId).let {
                 return@withContext Result.Success(it)
             }
-        }
-    }
-
-    /**
-     * Register new user and store the details.
-     *
-     * @param userEmail              email
-     * @param password               password
-     * @param firstName              first name
-     * @param lastName               last name
-     */
-    override suspend fun createNewUserAndStoreDetails(userEmail: String,
-                                                      password: String,
-                                                      firstName: String,
-                                                      lastName: String): Result<Boolean> {
-        return withContext(Dispatchers.IO) {
-            try {
-                userAuthentication.createNewUserWithEmailPassword(
-                        userEmail,
-                        password).await()
-                userAuthentication.currentUserUid?.let {
-                    onlineDatabaseStore.storeUserData(
-                            it,
-                            getStorableFromUserDetails(userEmail, firstName, lastName))
-                } ?: throw UnsupportedOperationException("User not logged in")
-            } catch (e: Exception) {
-                Result.Error(e)
-            }
-        }
-    }
-
-    private fun getStorableFromUserDetails(userEmail: String, firstName: String,
-                                           lastName: String): Storable {
-        return object : Storable {
-            override val hashMap: Map<String, Any>
-                get() = mapOf(
-                        "emailAddress" to userEmail,
-                        "firstName" to firstName,
-                        "lastName" to lastName
-                )
-
-            override val id: Long = 0
         }
     }
 
@@ -106,8 +62,10 @@ class MessageRepositoryImpl private constructor(
      *
      * @param message given Message
      */
-    override fun deleteMessage(message: Message) {
-        runOnThread { database.messageDao().deleteItem(message) }
+    override suspend fun deleteMessage(message: Message) {
+        withContext(Dispatchers.IO) {
+            database.messageDao().deleteItem(message)
+        }
     }
 
     /**
@@ -183,29 +141,21 @@ class MessageRepositoryImpl private constructor(
     }
 
     /**
-     * Call to logout user and clear all messages from Room.
-     */
-    override fun logOutUser() {
-        userAuthentication.signOutUser()
-        clearAllMessages()
-    }
-
-    /**
      * Call to clear all messages from Room.
      */
-    private fun clearAllMessages() {
-        database.messageDao().clear()
+    override suspend fun clearAllMessages() {
+        withContext(Dispatchers.IO) {
+            database.messageDao().clear()
+        }
     }
 
     /**
      * Add given list of messages to Room.
      */
-    private fun addMessagesToLocal(messages: List<Message>) {
-        runOnThread { database.messageDao().insertAllMessages(messages) }
-    }
-
-    private fun runOnThread(runnable: Runnable?) {
-        Thread(runnable).start()
+    private suspend fun addMessagesToLocal(messages: List<Message>) {
+        withContext(Dispatchers.IO) {
+            database.messageDao().insertAllMessages(messages)
+        }
     }
 
     /**
@@ -255,47 +205,6 @@ class MessageRepositoryImpl private constructor(
      */
     override suspend fun syncMessagesFromFireStoreToRoom() {
         onlineDatabaseStore.getAllMessagesForUser(userAuthentication.currentUserUid!!)
-    }
-
-    /**
-     * Get user name of the current user.
-     *
-     * @return user name
-     */
-    override val userName: String?
-        get() = userAuthentication.userName
-
-    /**
-     * Sign in user with email and password.
-     *
-     * @param email    email
-     * @param password password
-     * @param listener task completion listener`
-     */
-    override suspend fun signInWithEmailAndPassword(email: String, password: String): Result<Boolean> {
-        return withContext(Dispatchers.IO) {
-            try {
-                userAuthentication.signInWithEmailAndPassword(email, password).await().let {
-                    Result.Success(true)
-                }
-            } catch (e: Exception) {
-                Result.Error(e)
-            }
-        }
-    }
-
-    /**
-     * Reload the current user authentication state.
-     *
-     */
-    override suspend fun reloadCurrentUserAuthState(): Result<Boolean> {
-        return try {
-            userAuthentication.reloadCurrentUserAuthState()?.await()
-                    ?: throw Exception("User not logged in")
-            Result.Success(true)
-        } catch (e: Exception) {
-            Result.Error(e)
-        }
     }
 
     override suspend fun syncMessagesWithCloudAndLocal() {
