+++ app/src/main/java/com/fazemeright/chatbotmetcs622/ui/base/BaseViewModel.kt
@@ -5,11 +5,14 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import com.fazemeright.chatbotmetcs622.repositories.message.MessageRepository
 import com.fazemeright.chatbotmetcs622.repositories.message.MessageRepositoryImpl
+import com.fazemeright.chatbotmetcs622.repositories.user.UserRepository
+import com.fazemeright.chatbotmetcs622.repositories.user.UserRepositoryImpl
 import kotlinx.coroutines.launch
 
 abstract class BaseViewModel(application: Application) : AndroidViewModel(application) {
     @JvmField
     protected val messageRepository: MessageRepository = MessageRepositoryImpl.getInstance(application)
+    protected val userRepository: UserRepository = UserRepositoryImpl.getInstance(application)
 
     /**
      * Run on a UI safe thread.
@@ -26,13 +29,16 @@ abstract class BaseViewModel(application: Application) : AndroidViewModel(applic
      * @return username if logged in, else `null`
      */
     val userName: String?
-        get() = messageRepository.userName
+        get() = userRepository.userName
 
     /**
      * Logs out the user.
      */
     fun logOutUser() {
-        messageRepository.logOutUser()
+        userRepository.logOutUser()
+        viewModelScope.launch {
+            messageRepository.clearAllMessages()
+        }
     }
 
     fun syncMessagesWithLocalAndCloud() {
